Simplifies code
