+++ pansim/Cargo.toml
@@ -13,7 +13,6 @@ statrs = "0.16"
 rayon = "1.5"
 ndarray = { version = "0.15.6", features = ["rayon"] }
 clap = "3.0"
-hamming = "0.1"
 logsumexp = "0.1"
 
 [profile.release]
