+++ pansim/src/hamming.rs
@@ -1,0 +1,30 @@
+pub fn hamming_bitwise_fast(x: &[u8], y: &[u8]) -> u32 {
+    assert_eq!(x.len(), y.len());
+
+    // Process 8 bytes at a time using u64
+    let mut distance = x
+        .chunks_exact(8)
+        .zip(y.chunks_exact(8))
+        .map(|(x_chunk, y_chunk)| {
+            // This is safe because we know the chunks are exactly 8 bytes.
+            // Also, we don't care whether the platform uses little-endian or big-endian
+            // byte order. Since we're only XORing values, we just care that the
+            // endianness is the same for both.
+            let x_val = u64::from_ne_bytes(x_chunk.try_into().unwrap());
+            let y_val = u64::from_ne_bytes(y_chunk.try_into().unwrap());
+            (x_val ^ y_val).count_ones()
+        })
+        .sum::<u32>();
+
+    if x.len() % 8 != 0 {
+        distance += x
+            .chunks_exact(8)
+            .remainder()
+            .iter()
+            .zip(y.chunks_exact(8).remainder())
+            .map(|(x_byte, y_byte)| (x_byte ^ y_byte).count_ones())
+            .sum::<u32>();
+    }
+
+    distance
+}
