+++ pansim/src/lib.rs
@@ -1,0 +1,5 @@
+
+
+// Copyright 2024 Evan Schwartz from https://github.com/emschwartz/hamming-bitwise-fast/tree/main
+pub mod population;
+mod hamming;
